% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdffpRain-package.R
\docType{package}
\name{bdffpRain-package}
\alias{bdffpRain}
\alias{bdffpRain-package}
\title{bdffpRain: Rain gauge data from the Biological Dynamics of Forest Fragments project}
\description{
Provides daily precipitation records for 1987–2010 from eight rain guages across the Biological Dynamics of Forest Fragments Project (BDFFP), north of Manaus, Brazil.
}
\author{
\strong{Maintainer}: Eric Scott \email{scottericr@gmail.com} (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})

Authors:
\itemize{
  \item Emilio Bruna \email{embruna@ufl.edu} (\href{https://orcid.org/0000-0003-3381-8477}{ORCID})
}

}
\keyword{internal}
